backend/requirements.txt
# ── Core ─────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# ── Database ─────────────────────────────────────────
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.0.3

# ── Auth ─────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ── AI / ML ──────────────────────────────────────────
google-generativeai==0.7.2
openai==1.47.0
vaderSentiment==3.3.2
Pillow==10.4.0

# ── External APIs ────────────────────────────────────
tweepy==4.14.0
httpx==0.27.2

# ── Caching ──────────────────────────────────────────
redis==5.1.0

# ── Utilities ────────────────────────────────────────
python-dotenv==1.0.1

# ── Testing ──────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
